Port more main and grid functionality, clean up a bit
I can't say this commit completes the port of any one source file like
I'd hoped, though it does serve as significant progress for both grid
and main. It may have looked more focussed had I not been so
inconsistent in working on this project, though remedying that is the
point of this commit, anyway!

Essentially, the few changes I've made here to clean up the code are
indicative of my mindset moving forward. The main focus is to increase
readability, which was sorely lacking in prior commit messages as well
as a number of comments in the Python implementation. A major goal of
this project was proper structure, so I'm frankly surprised by the state
I left it in. There's still more to clean, though once complete I plan
on keeping that as the standard moving forward.