Split main.cpp into other src files, complete main.cpp
You could call this dishonest considering the main.cpp content of old is
not *technically* fully ported over, but I think it makes for better
project structure regardless.

Essentially, all the rendering code that used to be placed prior to the
main loop has been relocated to the new "render.cpp". As well, the
routine used to find the player's position in main has been moved to
grid.cpp. Neither of these functions have been fully ported over yet,
but the rest of main.cpp has. I think the former change makes the main
loop a lot cleaner while the latter makes sense given that said code is
likely to be reused in future.